+++ RELEASE_NOTES.md
@@ -22,6 +22,7 @@
 
 - Restored check-in on input screens (4.2.0)
 - Fixed default global background color (4.2.1)
+- Fixed display controller redirection on results or check-in input (4.2.1)
 
 ## _Sharly-Chess.com_